fix: CategoryRepository interface 로 구조 변경 - 다형성 활용 (for test)
