+++ src/main/java/com/wealdy/saemsembackend/domain/category/repository/CategoryRepository.java
@@ -1,35 +1,12 @@
 package com.wealdy.saemsembackend.domain.category.repository;
 
 import com.wealdy.saemsembackend.domain.category.entity.Category;
-import jakarta.persistence.EntityManager;
-import jakarta.persistence.NoResultException;
-import jakarta.persistence.TypedQuery;
 import java.util.List;
 import java.util.Optional;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Repository;
 
-@Repository
-@RequiredArgsConstructor
-public class CategoryRepository {
+public interface CategoryRepository {
 
-    private final EntityManager em;
+    List<Category> findAll();
 
-    public Long countAll() {
-        return em.createQuery("select count(c) from Category c", Long.class).getSingleResult();
-    }
-
-    public List<Category> findAll() {
-        return em.createQuery("select c from Category c", Category.class).getResultList();
-    }
-
-    public Optional<Category> findByName(String name) {
-        TypedQuery<Category> query = em.createQuery("select c from Category c where c.name=:name", Category.class)
-            .setParameter("name", name);
-        try {
-            return Optional.ofNullable(query.getSingleResult());
-        } catch (NoResultException e) {
-            return Optional.empty();
-        }
-    }
+    Optional<Category> findByName(String name);
 }
